requirements.txt
Pillow>=10.0.0
spacy>=3.7.0